requirements/base.txt
boto3
brotli
cachetools
falcon
honeybadger
lru-dict
orjson
psycopg[binary,pool]
pyparsing
requests
tenacity
testcontainers[postgres]
titlecase
zstandard
